switch to antd dropzone and update styling
