+++ src/Components/Atoms/DropBox.js
@@ -1,71 +1,60 @@
 import readFile from "../../Utils/readFile";
 import { callParse } from "../../Utils/callParse";
-import { useContext, useState } from "react";
-import { DataContext, useDataContext } from "../Contexts/DataContext";
+import { useState } from "react";
+import { useDataContext } from "../Contexts/DataContext";
+
+import { InboxOutlined } from '@ant-design/icons';
+import { message, Upload } from 'antd';
+import data from "../env/data";
+
+const { Dragger } = Upload;
 
 export function DropBox() {
+
+  
   // Dropbox component
   const [dragClass, setdragClass] = useState(false);
   const { setCpuData, setMemoryData, setBlockData, hasData } = useDataContext();
 
-  function handleDragOver(e) {
-    e.stopPropagation();
-    e.preventDefault();
-    e.dataTransfer.dropEffect = "copy";
+  const props = { // props for antd upload component
+    multiple: false,
+    customRequest: handleCustomRequest,
+    showUploadList: false,
   }
 
-  function handleDragEnter(e) {
-    // drag enter event to add classes/css
-    e.preventDefault();
-    setdragClass(true);
-    // document.getElementById('drop_zone').classList.add('dragging');
-    // document.getElementById('main-html').classList.add('darker');
-  }
+  async function handleCustomRequest({onError, onSuccess, file}) {
+    if(file) {
+      
+      const fileContent = await readFile(file);
+      const dataObj = callParse(fileContent); // Object containing more objects (inception! 🤯)
 
-  function handleDragLeave(e) {
-    // drag leave event to remove classes/css
-    e.preventDefault();
-    setdragClass(false);
-  }
+      // Save data in context
+      setCpuData(dataObj.cpuObject);
+      setMemoryData(dataObj.memoryObject);
+      setBlockData(dataObj.blockObject);
 
-  function handleClick(e) {
-    // Task #259
-    console.log(e);
-  }
-
-  async function handleFileSelect(e) {
-    // function to handle file selection and calling other functions
-    e.stopPropagation();
-    e.preventDefault();
-
-    let files = e.dataTransfer.files; // FileList object.
-    if (files.length > 1) {
-      // check if more than one file
-      alert("Only one file can be processed at a time.");
+      onSuccess()
     }
-    const fileContent = await readFile(files[0]);
-    const dataObj = callParse(fileContent); // Object containing more objects (inception! 🤯)
-
-    // Save data in context
-    setCpuData(dataObj.cpuObject);
-    setMemoryData(dataObj.memoryObject);
-    setBlockData(dataObj.blockObject);
   }
+
 
   function getStyles() {
-    return hasData ? "hidden" : "centered drop_zone";
+    return hasData ? "hidden" : "centered antd-dropzone";
   }
 
   return (
-    <div
-      className={`${dragClass ? "dragging" : ""} ${getStyles()}`}
-      onDragOver={handleDragOver}
-      onDragEnter={handleDragEnter}
-      onDragLeave={handleDragLeave}
-      onDrop={handleFileSelect}
-      onClick={handleClick}
-    >
-      Drop file here...
+    <div className={getStyles()}>
+      <Dragger {...props}>
+        <p className="ant-upload-drag-icon">
+          <InboxOutlined />
+        </p>
+        <p className="ant-upload-text">Click or drag file to this area to upload</p>
+        <p className="ant-upload-hint">
+
+          Select a sar file that contains the text output from systat (sa files are binary and won’t be read).
+          
+        </p>
+      </Dragger>
     </div>
   );
 }
