+++ src/Components/Styles/styles.css
@@ -39,32 +39,20 @@ a {
 }
 
 .centered {
-  top: 50%;
+  /* top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
   position: fixed;
-  margin: auto;
+  margin: auto; */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
 }
 
-.drop_zone {
-    background-color: rgb(107, 107, 107);
-    /* position: fixed; */
-    padding: 50px;
-    text-align: center;
-    border: 1px outset rgba(0, 70, 159, 0.32);
-    width: 40vw;
-    height: 30vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    /* top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%); */
-    -webkit-transition: background-color 500ms linear;
-    -ms-transition: background-color 500ms linear;
-    transition: background-color 500ms linear;
-    
+.antd-dropzone {
+  flex-grow: 1;
+  height: 100vh;
 }
 
 .drop_zone:hover {
